refactor(#3): 添加设备的代码全部移到了devices.py
